==> transcriptomics/Study1/2_cell_typing_and_nonMicroglia_removal.Rmd <==
---
title: "Cell Typing"
author: "David Tatarakis"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
  html_document:
    code_folding: hide
    toc: true
    toc_float: true
    toc_collapsed: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
libraries <- c("aws.s3", 
               "Seurat", 
               "Matrix", 
               "ggplot2", 
               "dplyr",
               "here")
lapply(libraries, library, character.only = T)
```

# Dimensionality reduction
```{r}
# read in the integrated object from the prior vignette
object <- readRDS("~/Study1_integrated.Rds")

# Make sure we're working off of the integrated RNA assay for these steps
DefaultAssay(object) <- "integrated"
object <- object %>%
  ScaleData(.) %>%
  RunPCA(.)

ElbowPlot(object)

dims <- 12

# Now we can run the dimentionality reductions
object <- object %>%
  RunUMAP(., dims = 1:dims) %>%
  FindNeighbors(., dims = 1:dims) %>%
  FindClusters(., resolution = 0.1)
``` 


```{r, fig.width=12, fig.height=4}
DimPlot(object, reduction = "umap", group.by = "batch") + DimPlot(object)
```


```{r}
DefaultAssay(object) <- "RNA"

# We need to normalize, find variable features, etc using this assay, now.
object <- object %>%
  NormalizeData(.)
```

```{r, eval = FALSE}
future::plan("multiprocess", workers = 4)
options(future.globals.maxSize = 12000 * 1024 ^ 2)

markers <- FindAllMarkers(object)

future::plan("sequential")

write.csv(markers,
          file = paste0(
            here(), 
            "/inst/extdata/Study1_cluster_markers.csv")
          )
```

```{r}
markers <-
  read.csv(
    paste0(
      here(), 
      "/inst/extdata/Study1_cluster_markers.csv"
      ),
    row.names = 1,
    header = T
  )

top10 <-
  markers %>% group_by(cluster) %>% slice_max(n = 10, order_by = avg_log2FC)
```

```{r, fig.width=6, fig.height=3}
FeaturePlot(object,
            c("P2ry12", 
              "Slc1a2", 
              "F13a1", 
              "Cldn5", 
              "S100a11", 
              "Atp1a2"),
            ncol = 3)
```

Time to remove the apparent non-microglia populations.
```{r, eval=FALSE}
microglia <- subset(object, subset = seurat_clusters %in% c(0, 1, 2))
saveRDS(microglia, file = "~/Study1_microglia.Rds")
```
